requirements-test.txt
pytest==6.2.1
ipython==7.19.0
pytest-aiohttp==0.3.0
